added character, behaviors, mines, factories, plants

+++ Assets/Game/Runtime/TileMap/Pathfinding/Heuristic.cs
@@ -4,7 +4,7 @@ namespace Game.Runtime.TileMap.Pathfinding
 {
     public class Heuristic : IHeuristic
     {
-        public float Value(Vector2Int first, Vector2Int second)
+        public double Value(Vector2Int first, Vector2Int second)
         {
             return System.Math.Abs(first.X - second.X) + System.Math.Abs(first.Y - second.Y);
         }
